Removidos archivos prefix (deprecated) y organización de los poise commands
