--- src/admin/ban/mod.rs
@@ -1,5 +1,0 @@
-pub mod ban;
-pub mod unban;
-
-pub use ban::*;
-pub use unban::*;
