--- src/admin/mod.rs
@@ -1,2 +1,0 @@
-pub mod ban;
-pub use ban::*;
