+++ src/fun/mod.rs
@@ -1,0 +1,2 @@
+pub mod action;
+pub use action::*;
