--- src/music_controller/mod.rs
@@ -1,21 +1,0 @@
-pub mod play;
-pub mod pause;
-pub mod resume;
-pub mod stop;
-pub mod mute;
-pub mod unmute;
-pub mod skip;
-pub mod utils;
-pub mod random;
-pub mod queue;
-
-pub use play::*;
-pub use pause::*;
-pub use resume::*;
-pub use stop::*;
-pub use mute::*;
-pub use unmute::*;
-pub use skip::*;
-pub use utils::*;
-pub use random::*;
-pub use queue::*;
