+++ src/utils/dependencies.rs
@@ -1,25 +1,47 @@
-pub use serenity::model::channel::Message;
-pub use serenity::{Result as SerenityResult};
-pub use std::sync::Arc;
-pub use anyhow::anyhow;
+pub use poise::serenity_prelude::{self as serenity_2, futures};
+pub use futures::{Stream, StreamExt};
+pub use serde::{Deserialize, Serialize};
+pub use crate::fun::*;
+pub use serenity::async_trait;
 
-pub use serenity::{
-    client::{Context},
-    prelude::Mutex,
-};
-pub use songbird::{
-    Call,
-};
+pub use crate::{Context, Error};
+pub use rand::{Rng, random};
 
-pub use serenity::framework::standard::{
-    macros::command,
-    CommandResult
-};
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Original {
+    pub url: String
+}
 
-pub use rand::seq::SliceRandom;
-pub use serenity::framework::standard::{Args};
-pub use songbird::input::Restartable;
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Data {
+    pub original: Original
+}
 
-pub use std::fs::read_to_string;
-pub use rand::random;
-pub use serenity::prelude::Mentionable;
+#[derive(Deserialize, Serialize, Debug)]
+pub struct Gif {
+    pub images: Data,
+    pub slug: String
+}
+
+#[derive(Deserialize, Serialize)]
+pub struct ResponseGiphy {
+    pub data: Vec<Gif>,
+}
+
+pub async fn autocomplete_actions<'a>(
+    _ctx: Context<'_>,
+    partial: &'a str,
+) -> impl Stream<Item = String> + 'a {
+    futures::stream::iter(&["greet", "kiss", "kick", "slap", "punch", "gun"])
+        .filter(move |name| futures::future::ready(name.starts_with(partial)))
+        .map(|name| name.to_string())
+}
+
+pub async fn autocomplete_status<'a>(
+    _ctx: Context<'_>,
+    partial: &'a str,
+) -> impl Stream<Item = String> + 'a {
+    futures::stream::iter(&["sleep", "playing", "boring", "angry", "happy" ])
+        .filter(move |name| futures::future::ready(name.starts_with(partial)))
+        .map(|name| name.to_string())
+}
